Create universal icon component (#347)
