--- src/components/icons/index.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-export const EvergreenLogo = () => (
-  <svg width="14" height="28" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M6.884 0C4.59 8.35 2.294 16.864 0 25.214c2.032-.246 3.934-.49 6.065-.818-.528 3.34 0 .235-.528 3.604h2.858l-.527-3.604c.164 0 2.95.49 5.737.982-.82-3.438-3.77-1.99-6.72-5.92 1.638.818 4.425 1.5 5.9 1.663-.983-3.438-3.278-1.5-5.9-4.938 1.31.655 3.605 1.172 4.917 1.5-.984-3.439-2.623-1.827-4.918-4.774 1.148.49 2.95 1.008 3.934 1.335-.82-3.438-1.803-2.154-3.934-4.774.984.492 2.131.845 2.95 1.172-.382-1.527-.621-2.063-.968-2.449-.396-.441-.933-.687-1.982-1.997.656.327 1.311.68 1.967.844-.82-2.947-.164-1.99-1.967-4.119.492.164.656.354.984.517L6.884 0z"
-      fill="#71CC97"
-      fillRule="evenodd"
-    />
-  </svg>
-);
