--- src/pages/userPatches/patchCard/BuildStatusIcon/FailedIcon.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-
-export const FailedIcon = () => (
-  <svg width="22px" height="22px" viewBox="0 0 22 22">
-    <g stroke="none" strokeWidth={1} fill="none" fillRule="evenodd">
-      <g transform="translate(-1124.000000, -648.000000)">
-        <g transform="translate(1125.000000, 649.000000)">
-          <circle stroke="#CF4A22" strokeWidth={2} cx={10} cy={10} r={10} />
-          <g transform="translate(4.000000, 4.000000)" fill="#CF4A22">
-            <path d="M6.55144413,0.144677043 C6.73102168,0.245689415 6.87941035,0.394078077 6.98042272,0.573655627 L11.8068443,9.15396066 C12.1114536,9.69548827 11.9193933,10.3814178 11.3778657,10.686027 C11.2094739,10.7807474 11.0195258,10.8305046 10.8263218,10.8305046 L1.17347866,10.8305046 C0.552158315,10.8305046 0.048478659,10.3268249 0.048478659,9.70550455 C0.048478659,9.51230059 0.0982357656,9.32235251 0.19295618,9.15396066 L5.01937776,0.573655627 C5.32398704,0.0321280151 6.00991652,-0.159932239 6.55144413,0.144677043 Z M5.99990024,7.77630533 C5.48213329,7.77630533 5.06240024,8.19603838 5.06240024,8.71380533 C5.06240024,9.23157229 5.48213329,9.65130533 5.99990024,9.65130533 C6.51766719,9.65130533 6.93740024,9.23157229 6.93740024,8.71380533 C6.93740024,8.19603838 6.51766719,7.77630533 5.99990024,7.77630533 Z M6.93740024,3.27630533 L5.06240024,3.27630533 L5.24990024,7.02630533 L6.74990024,7.02630533 L6.93740024,3.27630533 Z" />
-          </g>
-        </g>
-      </g>
-    </g>
-  </svg>
-);
