--- src/pages/userPatches/patchCard/BuildStatusIcon/RunningIcon.tsx
@@ -1,25 +1,0 @@
-import React from "react";
-
-export const RunningIcon = () => (
-  <svg width="22px" height="22px" viewBox="0 0 22 22">
-    <g stroke="none" strokeWidth={1} fill="none" fillRule="evenodd">
-      <g transform="translate(-1124.000000, -566.000000)">
-        <g id="Group-4" transform="translate(1125.000000, 567.000000)">
-          <circle
-            id="Oval"
-            stroke="#007CAD"
-            strokeWidth={2}
-            cx={10}
-            cy={10}
-            r={10}
-          />
-          <path
-            d="M10,17 C13.8659932,17 17,13.8659932 17,10 C17,6.13400675 13.8659932,3 10,3 C10,5.32759884 10,8.42828353 10,10 C3.27975701,11.9656673 10,10 3.27975701,11.9656673 C4.12925531,14.8747832 6.81644217,17 10,17 Z"
-            id="Oval"
-            fill="#007CAD"
-          />
-        </g>
-      </g>
-    </g>
-  </svg>
-);
