--- src/pages/userPatches/patchCard/BuildStatusIcon/SucceededIcon.tsx
@@ -1,23 +1,0 @@
-import React from "react";
-
-export const SucceededIcon = () => (
-  <svg width="22px" height="22px" viewBox="0 0 22 22">
-    <g stroke="none" strokeWidth={1} fill="none" fillRule="evenodd">
-      <g transform="translate(-1052.000000, -648.000000)">
-        <g id="Group-6" transform="translate(1053.000000, 649.000000)">
-          <g
-            id="Glyphs-/-Checkmark"
-            transform="translate(2.000000, 2.000000)"
-            fill="#13AA52"
-          >
-            <path
-              d="M13.3137085,4 L14.7279221,5.41421356 L6.94974747,13.1923882 L2,8.24264069 L3.41421356,6.82842712 L6.949,10.364 L13.3137085,4 Z"
-              id="Fill-/-All"
-            />
-          </g>
-          <circle stroke="#13AA52" strokeWidth={2} cx={10} cy={10} r={10} />
-        </g>
-      </g>
-    </g>
-  </svg>
-);
